build preprocessing data structure
The preprocessing script now builds a file that has in it a structure
containing separate training and validation datasets. The input filelist
is now only required to a be a text file where each row has the training
label (integer), sample name, and full filepath to the file to be slimmed.
This then allows for a data structure which can add attributes to the output
HDF5 groups that hold the training label, name, original file, and split
training and validation data.

By default the training and validation data
are split 50/50 based on the number of events in the sample with training label
0, which is assumed to be the signal. That is, if the signal has a total
of 10000 events, then for ALL samples the training set will have 5000 events
and the validation will have 5000 events. The user can manually set the
training size, but right now the script assumes that the validation
set will be the same size.

I need to think of a good way to combine all of the training/validation sets
into a super-dataset for training and validation purposes and to also build
the scaling quantities that are used to normalize each of the input fields once
the super dataset is built. The question is: do I do this at trianing,
and output a separate HDF5 file that has this scaling info per feature, or
do I determine this information at the preprocessing stage and simply
add another group to the output file that contains this information?