store scaling information for each field
the output file now contains a new group with a single dataset. the
dataset has the quantities needed for scaling the features for training,
namely the mean values, mean scale, and variances of each feature. these
values are determined using sklearn's StandardScaler and are calculated
using the total dataset composed of all input samples.