validation plots for NN output and log ratio
I now load the trained model and from this, after loading the pre-processed
files and validation set, I make histograms of the NN outputs for each of the
output labels of the network for each of the samples. Additionally, I histogram a log-ratio
discriminant for each of the output labels of the network for each of the samples.
Currently I am not weighting these histograms, this needs to be implemented.
Also, the next set of histograms will need to be the ROC curves.